requirements.txt
# Sigorta Otomasyon Projesi - Python Bağımlılıkları

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.1  # pymysql için gerekli

# Environment Variables
python-dotenv==1.0.1

# Web Scraping
playwright==1.48.0

# Authentication
pyotp==2.9.0

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Utilities
python-dateutil==2.9.0

# Logging
loguru==0.7.2

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1

